fix(v2.10/codegen): map Fin[N] handler params to u32 on Quasar (alignment-1)
Quasar's `#[instruction]` macro auto-converts native integer params
(`u64` → `PodU64`, …) inside the generated `InstructionDataZc`
struct. `usize` isn't in that map — it falls through and the ZC
struct fails the alignment-1 assertion. Resolution: emit `Fin[N]`
(and any alias like `AccountIdx = Fin[N]`) as `u32` in the lib.rs
shim signature so the auto-Pod conversion picks it up as `PodU32`.
The inner handler still takes `usize` for indexing; we cast at the
dispatch boundary with `arg as usize`.

This is the right shape because:
- WriteBytes is implemented for native integers (u8..u128), not for
  the Pod companions, so client-side `build_instruction_data` would
  fail if we exposed PodU32 directly in the user-facing signature.
- Quasar's macro already handles Pod-conversion of `u64`/`u128`/etc.
  internally; mirroring that for `usize` via a single `Fin → u32`
  rewrite keeps the shim idiomatic.

Drops percolator from 41 → 15 errors. The 13 alignment-1 cascades
plus the 14 "no method `.get` on usize" errors all clear in one
shot. Remaining: `PodU128 + PodI128` arithmetic mismatches in
guards (8 errors), one `mul_div_floor_u128` arg-type mismatch, three
`spec_hash = ""` qedgen-macro edge cases, and three method-on-Pod
issues.

Multisig and lending still compile cleanly. All 435 qedgen tests
pass.

+++ examples/rust/percolator/programs/src/lib.rs
@@ -17,33 +17,33 @@ mod percolator {
     use super::*;
 
     #[instruction(discriminator = 0)]
-    pub fn add_user(ctx: Ctx<AddUser>, i: usize) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i)
+    pub fn add_user(ctx: Ctx<AddUser>, i: u32) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize)
     }
 
     #[instruction(discriminator = 1)]
-    pub fn add_lp(ctx: Ctx<AddLp>, i: usize) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i)
+    pub fn add_lp(ctx: Ctx<AddLp>, i: u32) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize)
     }
 
     #[instruction(discriminator = 2)]
-    pub fn reclaim_empty_account(ctx: Ctx<ReclaimEmptyAccount>, i: usize) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i)
+    pub fn reclaim_empty_account(ctx: Ctx<ReclaimEmptyAccount>, i: u32) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize)
     }
 
     #[instruction(discriminator = 3)]
-    pub fn close_account(ctx: Ctx<CloseAccount>, i: usize) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i)
+    pub fn close_account(ctx: Ctx<CloseAccount>, i: u32) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize)
     }
 
     #[instruction(discriminator = 4)]
-    pub fn deposit(ctx: Ctx<Deposit>, i: usize, amount: u128) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i, amount)
+    pub fn deposit(ctx: Ctx<Deposit>, i: u32, amount: u128) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize, amount)
     }
 
     #[instruction(discriminator = 5)]
-    pub fn withdraw(ctx: Ctx<Withdraw>, i: usize, amount: u128) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i, amount)
+    pub fn withdraw(ctx: Ctx<Withdraw>, i: u32, amount: u128) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize, amount)
     }
 
     #[instruction(discriminator = 6)]
@@ -52,38 +52,38 @@ mod percolator {
     }
 
     #[instruction(discriminator = 7)]
-    pub fn deposit_fee_credits(ctx: Ctx<DepositFeeCredits>, i: usize, amount: u128) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i, amount)
+    pub fn deposit_fee_credits(ctx: Ctx<DepositFeeCredits>, i: u32, amount: u128) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize, amount)
     }
 
     #[instruction(discriminator = 8)]
-    pub fn convert_released_pnl(ctx: Ctx<ConvertReleasedPnl>, i: usize, x: u128) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i, x)
+    pub fn convert_released_pnl(ctx: Ctx<ConvertReleasedPnl>, i: u32, x: u128) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize, x)
     }
 
     #[instruction(discriminator = 9)]
-    pub fn execute_trade(ctx: Ctx<ExecuteTrade>, a: usize, b: usize, size_q: i128, exec_price: u64) -> Result<(), ProgramError> {
-        ctx.accounts.handler(a, b, size_q, exec_price)
+    pub fn execute_trade(ctx: Ctx<ExecuteTrade>, a: u32, b: u32, size_q: i128, exec_price: u64) -> Result<(), ProgramError> {
+        ctx.accounts.handler(a as usize, b as usize, size_q, exec_price)
     }
 
     #[instruction(discriminator = 10)]
-    pub fn liquidate_case_0(ctx: Ctx<LiquidateCase0>, i: usize) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i)
+    pub fn liquidate_case_0(ctx: Ctx<LiquidateCase0>, i: u32) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize)
     }
 
     #[instruction(discriminator = 11)]
-    pub fn liquidate_case_1(ctx: Ctx<LiquidateCase1>, i: usize) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i)
+    pub fn liquidate_case_1(ctx: Ctx<LiquidateCase1>, i: u32) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize)
     }
 
     #[instruction(discriminator = 12)]
-    pub fn liquidate_otherwise(ctx: Ctx<LiquidateOtherwise>, i: usize) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i)
+    pub fn liquidate_otherwise(ctx: Ctx<LiquidateOtherwise>, i: u32) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize)
     }
 
     #[instruction(discriminator = 13)]
-    pub fn settle_account(ctx: Ctx<SettleAccount>, i: usize) -> Result<(), ProgramError> {
-        ctx.accounts.handler(i)
+    pub fn settle_account(ctx: Ctx<SettleAccount>, i: u32) -> Result<(), ProgramError> {
+        ctx.accounts.handler(i as usize)
     }
 
     #[instruction(discriminator = 14)]
